Add "Wii U" Category.
Cemu added the ability to install base games to the mlc01 directory, mimicking the real Wii. These use title IDs to identify, so I've added Titles for each. Some may need to be modified to match the name on SteamGridDB.com more closely. Example: "Disney Epic Mickey 2: The Power of Two" is the proper title, but returns the wrong game in SGdb.com, so I renamed it to simply: "Disney Epic Mickey 2" to get the correct result.