Write icons before tags in steam shortcuts
While it is valid vdf, and steam will accept it, Steam itself always puts the tags
at the end of shortcuts.
This commit ensures that 3rd party tools can read the shortcut file.
It also ensures that the shortcuts are understandable by steam,
even if they make their own shortcut parser more strict in the future.